cockpit/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9
pytest>=7.4
pytest-django>=4.7
python-dotenv>=1.0
gunicorn>=21.2